feat: 대시보드 데이터 업데이트 함수 추가 -> utils
